+++ faster_coco_eval/extra/draw.py
@@ -27,7 +27,17 @@ def generate_ann_polygon(
     text: Optional[str] = None,
     legendgroup: Optional[str] = None,
     category_id_to_skeleton: Optional[dict] = None,
-):
+) -> go.Scatter:
+    """Generate annotation polygon for plotly.
+
+    ann: annotation dictionary
+    color: color of the annotation
+    iouType: type of the annotation bbox or segm or keypoints
+    text: text to display
+    legendgroup: legend group to display
+    category_id_to_skeleton: dictionary of category id to skeleton
+
+    """
     all_x = []
     all_y = []
 
@@ -96,7 +106,21 @@ def display_image(
     data_folder: Optional[str] = None,
     categories: Optional[list] = None,
     return_fig: bool = False,
-):
+) -> Optional[go.Figure]:
+    """
+    Display the image with the results
+    cocoGt: ground truth
+    cocoDt: detection
+    image_id: image id
+    iouType: type of the annotation bbox or segm or keypoints
+    display_fp: display false positive
+    display_fn: display false negative
+    display_tp: display true positive
+    display_gt: display ground truth
+    data_folder: data folder
+    categories: categories to display
+    return_fig: return the figure
+    """
     polygons = []
 
     image = cocoGt.imgs[image_id]
@@ -253,7 +277,14 @@ def display_matrix(
     labels: list,
     normalize: bool = False,
     return_fig: bool = False,
-):
+) -> Optional[go.Figure]:
+    """
+    Display the confusion matrix
+    conf_matrix: confusion matrix
+    labels: labels of the matrix
+    normalize: normalize the matrix
+    return_fig: return the figure
+    """
     _labels = labels + ["fp", "fn"]
 
     if normalize:
@@ -316,6 +347,11 @@ def display_matrix(
 
 
 def plot_pre_rec(curves, return_fig: bool = False):
+    """
+    Plot the precision-recall curve
+    curves: list of curves to plot
+    return_fig: return the figure
+    """
     fig = go.Figure()
 
     for _curve in curves:
@@ -364,11 +400,16 @@ def plot_pre_rec(curves, return_fig: bool = False):
 
 
 def plot_f1_confidence(curves, return_fig: bool = False):
+    """
+    Plot the F1 confidence curve
+    curves: list of curves to plot
+    return_fig: return the figure
+    """
     fig = go.Figure()
     eps = 1e-16
     for _curve in curves:
         recall_list = _curve["recall_list"]
-        precision_list = _curve["precision_list"]
+        precision_list = _curve["precision_list"][: len(recall_list)]
         scores = _curve["scores"]
         f1_curve = (
             2
@@ -406,6 +447,82 @@ def plot_f1_confidence(curves, return_fig: bool = False):
         "autosize": True,
         "height": 600,
         "width": 1200,
+    }
+
+    fig.update_layout(layout)
+
+    if return_fig:
+        return fig
+
+    fig.show()
+
+
+def plot_ced_metric(curves, normalize: bool = False, return_fig: bool = False):
+    fig = go.Figure()
+
+    if normalize:
+        fig.layout.yaxis.title = (
+            "The proportion of the sample to the total sample [%]"
+        )
+        _hovertemplate_y = "%{y:.2f}%<br>"
+    else:
+        fig.layout.yaxis.title = "Number of samples"
+        _hovertemplate_y = "n=%{y}<br>"
+
+    for ced_curve in curves:
+        if normalize:
+            y = (np.array(ced_curve["count"]) / ced_curve["total_count"]) * 100
+        else:
+            y = ced_curve["count"]
+
+        fig.layout.xaxis.title = "Mean squared error"
+
+        category_name = ced_curve["category"]["name"]
+        fig.add_trace(
+            go.Scatter(
+                x=ced_curve["mse"],
+                y=y,
+                name=f"CED Curve [{category_name}]",
+                hovertemplate=_hovertemplate_y
+                + "mse: %{x:.2f}<br><extra></extra>",
+                showlegend=True,
+                mode="lines",
+            )
+        )
+
+    fig.update_xaxes(showspikes=True)
+    fig.update_yaxes(showspikes=True)
+
+    updatemenus = [
+        dict(
+            type="dropdown",
+            direction="down",
+            y=1.1,
+            x=1,
+            buttons=list(
+                [
+                    dict(
+                        args=[{"xaxis.type": "linear"}],
+                        label="Linear Scale",
+                        method="relayout",
+                    ),
+                    dict(
+                        args=[{"xaxis.type": "log"}],
+                        label="Log Scale",
+                        method="relayout",
+                    ),
+                ]
+            ),
+        ),
+    ]
+
+    layout = {
+        "title": "Cumulative Error Distribution",
+        "autosize": True,
+        "height": 600,
+        "width": 1200,
+        # "xaxis_type" : "log",
+        "updatemenus": updatemenus,
     }
 
     fig.update_layout(layout)